+++ crates/wasm/src/lib.rs
@@ -1,4 +1,6 @@
+extern crate console_error_panic_hook;
 
+use std::panic;
 use std::str::{from_utf8, from_utf8_unchecked};
 use std::io::Cursor;
 
@@ -34,7 +36,10 @@ use moclib::moc2d::{
 use moclib::deser::{
   ascii::{from_ascii_ivoa, moc2d_from_ascii_ivoa},
   json::{from_json_aladin, cellmoc2d_from_json_aladin},
-  fits::{from_fits_ivoa, MocIdxType}
+  fits::{
+    from_fits_ivoa, MocIdxType,
+    multiordermap::from_fits_multiordermap
+  }
 };
 
 pub(crate) mod common;
@@ -62,6 +67,12 @@ const JD_TO_USEC: f64 = (24_u64 * 60 * 60 * 1_000_000) as f64;
 extern "C" {
   #[wasm_bindgen(js_namespace = console)]
   fn log(s: &str);
+}
+
+/// Activate debugging mode (Rust stacktrace)
+#[wasm_bindgen(js_name = "debugOn")]
+pub fn debug_on() {
+  console_error_panic_hook::set_once();
 }
 
 /////////////////////////////
@@ -253,6 +264,82 @@ pub fn from_local_file(qtype: Option<MocQType>) -> Result<(), JsValue> {
   Ok(())
 }
 
+/// Open the file selection dialog and load the mulit-order-map the fits file contains 
+/// (for security reasons, we cannot simply provide a path on the client machine).
+/// # Warning
+/// Because of security restriction, the call to this method
+/// **"needs to be triggered within a code block that was the handler of a user-initiated event"**
+#[wasm_bindgen(js_name = "fromLocalMultiOrderMap", catch)]
+pub fn from_local_multiordermap(
+  from_threshold: f64,
+  to_threshold: f64,
+  asc: bool,
+  not_strict: bool,
+  split: bool,
+  revese_recursive_descent: bool,
+) -> Result<(), JsValue> {
+  // Create the file input action that will be fired by the event 'change'
+  let file_input_action = Closure::wrap(Box::new(move |event: Event| {
+    let element = unsafe { event.target().unchecked_unwrap().dyn_into::<HtmlInputElement>().unchecked_unwrap_ok() };
+    let filelist = unsafe {  element.files().unchecked_unwrap() };
+    for i in 0..filelist.length() {
+      let file = unsafe {  filelist.get(i).unchecked_unwrap() };
+      let file_name = file.name();
+      let file_reader = unsafe {  FileReader::new().unchecked_unwrap_ok() };
+      // There is a stream method, but I am not sure how to use it. I am so far going the easy way.
+      match file_reader.read_as_array_buffer(&file) {
+        Err(_) => log("Error reading file content"),
+        _ => { },
+      };
+      let file_onload = Closure::wrap(Box::new(move |event: Event| {
+        let file_reader: FileReader = unsafe { event.target().unchecked_unwrap().dyn_into().unchecked_unwrap_ok() };
+        let file_content = unsafe { file_reader.result().unchecked_unwrap_ok() };
+        let file_content: Vec<u8> = js_sys::Uint8Array::new(&file_content).to_vec();
+        // log(&format!("File len {:?}", file_content.len()));
+        // We accept only ".fits" files so splitting on "." should be safe.
+        let (name, ext) = unsafe { file_name.rsplit_once('.').unchecked_unwrap() };
+        let res = match ext {
+          "fits" => from_multitordermap_fits_file(
+            name, &file_content, from_threshold, to_threshold, 
+            asc, not_strict, split, revese_recursive_descent
+          ),
+          _ => unreachable!(), // since file_input.set_attribute("accept", ".fits");
+        };
+        match res {
+          Err(e) => log(&e.as_string().unwrap_or_else(|| String::from("Error parsing file"))),
+          _ => { },
+        };
+      }) as Box<dyn FnMut(_)>);
+      file_reader.set_onload(Some(file_onload.as_ref().unchecked_ref()));
+      file_onload.forget();
+    }
+  }) as Box<dyn FnMut(_)>);
+
+  // Create a temporary input file and click on it
+  // - get the body
+  let window = web_sys::window().expect("no global `window` exists");
+  // This could be used but not yet in web_sys: https://developer.mozilla.org/en-US/docs/Web/API/Window/showOpenFilePicker
+  let document = window.document().expect("should have a document on window");
+  let body = document.body().expect("document should have a body");
+  // - create the input
+  let file_input: HtmlInputElement = unsafe { document.create_element("input").unchecked_unwrap_ok().dyn_into()? };
+  file_input.set_type("file");
+  unsafe {
+    file_input.set_attribute("multiple", "").unchecked_unwrap_ok();
+    file_input.set_attribute("hidden", "").unchecked_unwrap_ok();
+    file_input.set_attribute("accept", ".fits").unchecked_unwrap_ok();
+  }
+  file_input.add_event_listener_with_callback("change", file_input_action.as_ref().unchecked_ref())?;
+  file_input_action.forget();
+  // - attach the input
+  body.append_child(&file_input)?;
+  // - simulate a click
+  file_input.click();
+  // - remove the input
+  body.remove_child(&file_input)?;
+  Ok(())
+}
+
 // - from FITS 
 
 #[wasm_bindgen(js_name = "fromFits", catch)]
@@ -268,13 +355,99 @@ pub fn from_fits(name: &str, data: &[u8]) -> Result<(), JsValue> {
   store::add(name, moc)
 }
 
-/// WARNING: if this i not working, check e.g. with `wget -v -S ${url}` the the content type is
+/// Create o S-MOC from a FITS multi-prder map plus other parameters.
+/// * `from_threshold`: Cumulative value at which we start putting cells in he MOC (often = 0).
+/// * `to_threshold`: Cumulative value at which we stop putting cells in the MOC.
+/// * `asc`: Compute cumulative value from ascending density values instead of descending (often = false).
+/// * `not_strict`: Cells overlapping with the upper or the lower cumulative bounds are not rejected (often = false).
+/// * `split`: Split recursively the cells overlapping the upper or the lower cumulative bounds (often = false).
+/// * `revese_recursive_descent`: Perform the recursive descent from the highest to the lowest sub-cell, only with option 'split' (set both flags to be compatibile with Aladin)
+#[wasm_bindgen(js_name = "fromFitsMulitOrderMap", catch)]
+pub fn from_multitordermap_fits_file(
+  name: &str, 
+  data: &[u8],
+  from_threshold: f64,
+  to_threshold: f64,
+  asc: bool,
+  not_strict: bool,
+  split: bool,
+  revese_recursive_descent: bool,
+) -> Result<(), JsValue> {
+  let moc = from_fits_multiordermap(
+    Cursor::new(data),
+    from_threshold,
+    to_threshold,
+    asc,
+    !not_strict,
+    split,
+    revese_recursive_descent
+  ).map_err(|e| JsValue::from_str(&e.to_string()))?;
+  // Add it to the store
+  store::add(name, InternalMoc::Space(moc))
+}
+
+/*
+#[wasm_bindgen(js_name = "fromFitsMulitOrderMapStd", catch)]
+pub fn from_mutlitordermap_fits_file_std(
+  name: &str,
+  data: &[u8],
+) -> Result<(), JsValue> {
+  let moc = from_fits_multiordermap(
+    Cursor::new(data),
+    0.0,
+    0.9,
+    false,
+    true,
+    false,
+    false
+  ).map_err(|e| JsValue::from_str(&e.to_string()))?;
+  // Add it to the store
+  store::add(name, InternalMoc::Space(moc))
+}*/
+
+/// WARNING: if this is not working, check e.g. with `wget -v -S ${url}` the the content type is
 /// `Content-Type: application/fits`.
 #[wasm_bindgen(js_name = "fromFitsUrl")]
 pub async fn from_fits_url(name: String, url: String) -> Result<(), JsValue> {
   from_url(name, url, "application/fits", Box::new(from_fits)).await
 }
 
+
+/// WARNING: if this is not working, check e.g. with `wget -v -S ${url}` the the content type is
+/// `Content-Type: application/fits`.
+ #[wasm_bindgen(js_name = "fromMultiOrderMapFitsUrl")]
+pub async fn from_multiordermap_url(
+  name: String,
+  url: String,
+  from_threshold: f64,
+  to_threshold: f64,
+  asc: bool,
+  not_strict: bool,
+  split: bool,
+  revese_recursive_descent: bool,
+) -> Result<(), JsValue>
+{
+  let func = move |name: &str, data: &[u8]| from_multitordermap_fits_file(
+    name, 
+    data,
+    from_threshold,
+    to_threshold,
+    asc,
+    not_strict,
+    split,
+    revese_recursive_descent
+  );
+  from_url(name, url, "application/fits", Box::new(func)).await
+}
+
+/*
+/// WARNING: if this is not working, check e.g. with `wget -v -S ${url}` the the content type is
+/// `Content-Type: application/fits`.
+#[wasm_bindgen(js_name = "fromMultiOrderMapFitsUrlStd")]
+pub async fn from_multiordermap_url_std(name: String, url: String) -> Result<(), JsValue> {
+  console_error_panic_hook::set_once();
+  from_url(name, url, "application/fits", Box::new(from_mutlitordermap_fits_file_std)).await
+}*/
 
 // - from ASCII