+++ crates/wasm/CHANGELOG.md
@@ -6,6 +6,9 @@ Released 2021-xx-xx.
 
 ### Added
 
+* `from_local_multiordermap`: load a multi-order map from a local file and create a MOC
+* `from_multiordermap_url`: load a multi-order map from a URL file and create a MOC
+* `from_multiordermap_fits_file`:  parse a multi-order map FITS file content and create a MOC
 * `from_valued_cells`: MOC from multi-order map
 * `split`: split a disjoint S-MOC into joint S-MOCs
 * `split_count`: count the number of joint S-MOCs in a disjoint S-MOC